Add free_invoice wizard in a separate file
